requirements.txt
# Arquivo: requirements.txt | Objetivo: dependencias Python necessarias para executar o sistema.
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==2.3.7
pandas==3.0.0
openpyxl>=3.1.5
xlrd==2.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.9
